+++ tox.ini
@@ -61,7 +61,7 @@ commands =
 
 [testenv:cybertracker]
 deps =
-    -r{toxinidir}/f/connectors/cybertracker/cybertracker_observations.script.lock
+    -r{toxinidir}/f/connectors/cybertracker/cybertracker_observations_from_backup.script.lock
     -r{toxinidir}/f/connectors/cybertracker/tests/requirements-test.txt
 commands =
     pytest {posargs} f/connectors/cybertracker
